requirements.txt
albumentations==1.3.1
clearml==1.12.20
click==8.1.7
#dvc[gdrive]==3.21.1
dvc==3.51.2
dvclive[lightning]==2.16.0
GitPython==3.1.36
jpeg4py==0.1.4
loguru==0.7.2
numpy==1.25.2
omegaconf==2.3.0
opencv-python==4.8.0.76
pandas==2.1.0
pydantic==2.1.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
pytorch-lightning[extra]==2.0.8
timm==0.9.7
torch==2.0.1
torchvision==0.15.2
